pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "sobamchan", email = "oh.sore.sore.soutarou@gmail.com"},
]
dependencies = [
    "discord-py>=2.3.2",
    "requests>=2.31.0",
    "schnitsum>=0.4.10",
    "python-dotenv>=1.0.0",
    "feedparser>=6.0.10",
    "sienna>=0.2.4",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "ipython>=8.16.1",
    "neovim>=0.3.1",
]
